+++ src/components/ColorPicker/modal.tsx
@@ -1,12 +1,15 @@
 'use client';
 
 import React, { useEffect, useState } from 'react';
+
+import tinycolor from 'tinycolor2';
 import Wheel from '@uiw/react-color-wheel';
 
 import { CursorPreview } from './preview';
-import { CloseSVG, RefreshSVG } from '@components/svgs';
+import { CloseSVG, RefreshSVG, PaletteSVG } from '@components/svgs';
 
 import { generateRandomColors } from '@utils/randomColors';
+import { WATCH_COLORS } from '@root/configs';
 
 import { Color } from 'bibata/app';
 
@@ -27,7 +30,9 @@ const ColorWheelCard: React.FC<ColorWheelCardProps> = (props) => {
 
   return (
     <div className='flex flex-col justify-center items-center'>
-      <div className='font-bold text-xs sm:text-md'>{props.title}</div>
+      <div className='font-bold text-xs sm:text-md text-center'>
+        {props.title}
+      </div>
       <div className='text-xl sm:text mt-2 p-2'>{props.children}</div>
       <input
         type='text'
@@ -68,16 +73,48 @@ export const ColorPickerModal: React.FC<ColorPickerModalProps> = (props) => {
     }
   });
 
+  const [monochromeMode, setMonochromeMode] = useState(false);
+
   const defaultColor = generateRandomColors();
-  const [baseColor, setBaseColor] = useState<string>(defaultColor[0]);
-  const [outlineColor, setOutlineColor] = useState<string>(defaultColor[1]);
-  const [watchColor, setWatchColor] = useState<string>(defaultColor[2]);
+  const [baseColor, setBaseColor] = useState(defaultColor[0]);
+  const [outlineColor, setOutlineColor] = useState(defaultColor[1]);
+  const [watchBGColor, setWatchBGColor] = useState(defaultColor[2]);
+  const [watchColor1, setWatchColor1] = useState(WATCH_COLORS.c1);
+  const [watchColor2, setWatchColor2] = useState(WATCH_COLORS.c2);
+  const [watchColor3, setWatchColor3] = useState(WATCH_COLORS.c3);
+  const [watchColor4, setWatchColor4] = useState(WATCH_COLORS.c4);
+
+  const watchColor = {
+    bg: watchBGColor,
+    c1: watchColor1,
+    c2: watchColor2,
+    c3: watchColor3,
+    c4: watchColor4
+  };
 
   const refreshColors = () => {
     const color = generateRandomColors();
     setBaseColor(color[0]);
     setOutlineColor(color[1]);
-    setWatchColor(color[2]);
+    setWatchBGColor(color[2]);
+
+    monoWedgeColors(color[0], color[2]);
+  };
+
+  const monoWedgeColors = (b: string = baseColor, w: string = watchBGColor) => {
+    if (monochromeMode) {
+      const pallete = tinycolor.mix(b, w, 10).monochromatic(4);
+      const colors = pallete.map((p) => p.toHexString());
+      setWatchColor1(colors[1]);
+      setWatchColor2(colors[2]);
+      setWatchColor3(colors[3]);
+      setWatchColor4(colors[0]);
+    } else {
+      setWatchColor1(WATCH_COLORS.c1);
+      setWatchColor2(WATCH_COLORS.c2);
+      setWatchColor3(WATCH_COLORS.c3);
+      setWatchColor4(WATCH_COLORS.c4);
+    }
   };
 
   const handleColorPick = () => {
@@ -104,33 +141,54 @@ export const ColorPickerModal: React.FC<ColorPickerModalProps> = (props) => {
     }
   }, []);
 
+  useEffect(() => {
+    monoWedgeColors();
+  }, [monochromeMode, baseColor]); // eslint-disable-line react-hooks/exhaustive-deps
+
   return (
     <>
       {props.isOpen && (
         <div
           onClick={(e) => e.target === e.currentTarget && props.onClose()}
           className='z-20 fixed top-0 left-0 w-full h-full flex items-center justify-center bg-black backdrop-filter backdrop-blur-xl firefox:bg-opacity-80 bg-opacity-80'>
-          <div className='w-full md:w-1/2 xl:w-1/3 max-h-full overflow-y-auto p-4 m-4 rounded-3xl shadow-lg bg-black ring-1 ring-white/[.06]'>
+          <div
+            className='w-full md:w-1/2 xl:w-1/3 max-h-full overflow-y-auto p-4 sm:m-4 sm:rounded-3xl sm:shadow-lg ring-1 ring-white/[.06]'
+            style={{
+              backgroundColor: `color-mix(in srgb, #161616 98%, ${baseColor})`
+            }}>
             <div className='flex justify-between text-xs'>
+              <div className='inline-flex gap-3'>
+                <button
+                  className='p-2 bg-white/[.1] text-white rounded-lg sm:rounded-2xl hover:bg-green-400 active:bg-green-200 hover:text-black'
+                  onClick={refreshColors}>
+                  <RefreshSVG />
+                </button>
+
+                <button
+                  className={`py-2 px-5 rounded-lg sm:rounded-2xl text-xs sm:text-lg font-bold ${
+                    monochromeMode
+                      ? 'bg-white/[.2] hover:bg-white/[.6] hover:text-black'
+                      : 'bg-white hover:bg-white/[.7] text-black'
+                  }`}
+                  onClick={() => setMonochromeMode((b) => !b)}>
+                  <PaletteSVG />
+                </button>
+              </div>
+
               <button
-                className='p-2 bg-[#93f5d2] text-black rounded-2xl hover:bg-[#d9c57f] active:bg-[#ffc68f]'
-                onClick={refreshColors}>
-                <RefreshSVG />
-              </button>
-              <button
-                className='p-3 bg-gray-500 text-white rounded-2xl font-bold hover:bg-gray-400'
+                className='p-3 bg-white/[.1] text-white rounded-xl font-bold hover:bg-white hover:text-black'
                 onClick={props.onClose}>
                 <CloseSVG />
               </button>
             </div>
-            <div className='h-48 sm:h-96 md:h-72 mt-2'>
+            <div className='mt-5 p-5 h-48 sm:h-96 md:h-72'>
               <CursorPreview
                 base={baseColor}
                 outline={outlineColor}
                 watch={watchColor}
               />
             </div>
-            <div className='mt-8 grid grid-cols-3 gap-10'>
+            <div className='mt-3 grid grid-cols-3 gap-10'>
               <ColorWheelCard
                 title='Base'
                 value={baseColor}
@@ -160,19 +218,84 @@ export const ColorPickerModal: React.FC<ColorPickerModalProps> = (props) => {
               </ColorWheelCard>
 
               <ColorWheelCard
-                title='Watch BG'
-                value={watchColor}
-                onChange={(c) => setWatchColor(c)}>
+                title='Watch Background'
+                value={watchBGColor}
+                onChange={(c) => setWatchBGColor(c)}>
                 <Wheel
                   width={wheelSize}
                   height={wheelSize}
-                  color={watchColor}
+                  color={watchBGColor}
                   onChange={(cr) => {
-                    setWatchColor(cr.hex);
+                    setWatchBGColor(cr.hex);
                   }}
                 />
               </ColorWheelCard>
+
+              {!monochromeMode && (
+                <>
+                  <ColorWheelCard
+                    title='Watch Color 1'
+                    value={watchColor1}
+                    onChange={(c) => setWatchColor1(c)}>
+                    <Wheel
+                      width={wheelSize}
+                      height={wheelSize}
+                      color={watchColor1}
+                      onChange={(cr) => {
+                        setWatchColor1(cr.hex);
+                      }}
+                    />
+                  </ColorWheelCard>
+
+                  <ColorWheelCard
+                    title='Watch Color 2'
+                    value={watchColor2}
+                    onChange={(c) => setWatchColor2(c)}>
+                    <Wheel
+                      width={wheelSize}
+                      height={wheelSize}
+                      color={watchColor2}
+                      onChange={(cr) => {
+                        setWatchColor2(cr.hex);
+                      }}
+                    />
+                  </ColorWheelCard>
+
+                  <ColorWheelCard
+                    title='Watch Color 3'
+                    value={watchColor3}
+                    onChange={(c) => setWatchColor3(c)}>
+                    <Wheel
+                      width={wheelSize}
+                      height={wheelSize}
+                      color={watchColor3}
+                      onChange={(cr) => {
+                        setWatchColor3(cr.hex);
+                      }}
+                    />
+                  </ColorWheelCard>
+
+                  <span />
+
+                  <ColorWheelCard
+                    title='Watch Color 4'
+                    value={watchColor4}
+                    onChange={(c) => setWatchColor4(c)}>
+                    <Wheel
+                      width={wheelSize}
+                      height={wheelSize}
+                      color={watchColor4}
+                      onChange={(cr) => {
+                        setWatchColor4(cr.hex);
+                      }}
+                    />
+                  </ColorWheelCard>
+
+                  <span />
+                </>
+              )}
             </div>
+
             <div className='mt-11 flex justify-center'>
               <button
                 className='w-36 py-3 bg-green-600 text-white font-bold rounded-2xl text-sm sm:text-md hover:bg-green-500'