+++ src/configs.ts
@@ -1,6 +1,6 @@
-import { Colors, Delays } from 'bibata/app';
+import { Color, Colors, Delays } from 'bibata/app';
 
-export const VERSIONS = ['1.0.0-alpha.0', '1.0.0-alpha.1'];
+export const VERSIONS = ['1.0.0-alpha.0', '1.0.0-alpha.1', '1.0.0-beta.0'];
 
 export const SPONSOR_API_ENDPOINT =
   'https://sponsor-spotlight.vercel.app/api/fetch';
@@ -13,14 +13,21 @@ export const TYPES = ['Modern', 'Original'];
 
 export const SIZES = [16, 20, 22, 24, 28, 32, 40, 48, 56, 64, 72, 80, 88, 96];
 
-export const COLORS_MASK = {
+export const WATCH_COLORS = {
+  c1: '#32a0da',
+  c2: '#7eba41',
+  c3: '#f05024',
+  c4: '#fcb813'
+};
+
+export const COLORS_MASK: Color = {
   base: '#00ff00',
   outline: '#0000ff',
-  watch: '#ff0000'
+  watch: { bg: '#ff0000', ...WATCH_COLORS }
 };
 
 export const COLORS: Colors = {
-  Amber: { base: '#ff8300', outline: '#ffffff', watch: '#001524' },
+  Amber: { base: '#ff8300', outline: '#ffffff', watch: { bg: '#001524' } },
   Classic: { base: '#000000', outline: '#ffffff' },
   Ice: { base: '#ffffff', outline: '#000000' }
 };
@@ -33,6 +40,6 @@ export const DELAYS: Delays = {
 
 export const DB_SEEDS = {
   FRESH_SIGNUP_DOWNLOADS: 20,
-  DOWNLOAD_MULTIPLIER: 10,
-  DOWNLOADS_PER_CENTS: (cents: number) => cents * DB_SEEDS.DOWNLOAD_MULTIPLIER // 1000 public Downloads per dollar
+  DOWNLOAD_MULTIPLIER: 3,
+  DOWNLOADS_PER_CENTS: (cents: number) => cents * DB_SEEDS.DOWNLOAD_MULTIPLIER // 300 public Downloads per dollar
 };